Feat : ItemRemover

+++ Assets/Scripts/GameManager.cs
@@ -9,7 +9,7 @@ public class GameManager : MonoBehaviour
     public bool IsPlaying { get; private set; } //게임 진행 여부
     public int score { get; private set; } //점수
 
-    public int groundSpeed { get; private set; }
+    public float groundSpeed { get; private set; }
 
     private void Awake()
     {
@@ -39,7 +39,7 @@ public class GameManager : MonoBehaviour
     {
         IsPlaying = false;
     }
-    public void SetGroundSpeed(int speed)
+    public void SetGroundSpeed(float speed)
     {
         groundSpeed += speed;
     }
